requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.23
botocore==1.19.23
CacheControl==0.12.6
certifi==2020.11.8
chardet==3.0.4
idna==2.10
idna-ssl==1.1.0
iso3166==1.0.1
jmespath==0.10.0
mapbox==0.18.0
msgpack==1.0.0
multidict==5.0.2
polyline==1.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
websockets==8.1
yarl==1.6.3
